task_2/chatbot_project/requirements.txt
streamlit
openai
qdrant-client
pandas
numpy
python-dotenv
sentence-transformers
langchain
langchain-openai
scikit-learn
pinecone
scipy
matplotlib
plotly